+++ sslutils.h
@@ -57,8 +57,8 @@ extern char ssl_rootfile[PATH_MAX];
 void init_ssl_or_die(int is_server, int enable_compression);
 SSL *ssl_promotion(int is_server, SOCKET s, char *errbuf, size_t errbuflen);
 SSL *ssl_promotion_rw(int is_server, SOCKET in, SOCKET out, char *errbuf, size_t errbuflen);
-int ssl_send(SSL *, char const *buffer, size_t size, char *errbuf, size_t errbuflen);
-int ssl_recv(SSL *, char *buffer, size_t size, char *errbuf, size_t errbuflen);
+int ssl_send(SSL *, char const *buffer, int size, char *errbuf, size_t errbuflen);
+int ssl_recv(SSL *, char *buffer, int size, char *errbuf, size_t errbuflen);
 
 #else   // HAVE_OPENSSL
